Code review changes
- Add Davies-Bouldin to clustering test_common
- Fix flake8 issue
- Use bare `assert` in test
- Fix typo

+++ sklearn/metrics/cluster/unsupervised.py
@@ -297,7 +297,7 @@ def davies_bouldin_index(X, labels):
     check_number_of_labels(n_labels, n_samples)
 
     intra_dists = np.zeros(n_labels)
-    centroids = np.zeros((n_labels, len(X[0])), dtype=X.dtype)
+    centroids = np.zeros((n_labels, len(X[0])), dtype=np.float)
     for k in range(n_labels):
         cluster_k = safe_indexing(X, labels == k)
         centroid = cluster_k.mean(axis=0)
